+++ src/common/config_parser.cpp
@@ -273,6 +273,8 @@ void ConfigParser::addOptionsModel(po::options_description& desc) {
       extra-output")
     ("trainable-interpolation", po::value<bool>()->zero_tokens()->default_value(false),
      "Enable trainable lm for the lm interpolation models.")
+    ("lm-cost", po::value<double>()->default_value(0.0),
+     "Add a training cost with an LM.")
     ("dim-vocabs", po::value<std::vector<int>>()
       ->multitoken()
       ->default_value(std::vector<int>({0, 0}), "0 0"),
@@ -399,7 +401,9 @@ void ConfigParser::addOptionsTraining(po::options_description& desc) {
     ("no-shuffle", po::value<bool>()->zero_tokens()->default_value(false),
     "Skip shuffling of training data before each epoch")
     ("tempdir,T", po::value<std::string>()->default_value("/tmp"),
-      "Directory for temporary (shuffled) files")
+      "Directory for temporary (shuffled) files and database")
+    ("sqlite", po::value<bool>()->zero_tokens()->default_value(false),
+      "Use temporary disk-based sqlite3 database for training corpus storage")
     ("devices,d", po::value<std::vector<int>>()
       ->multitoken()
       ->default_value(std::vector<int>({0}), "0"),
@@ -783,7 +787,9 @@ void ConfigParser::parseOptions(int argc, char** argv, bool doValidate) {
     SET_OPTION("save-freq", size_t);
     SET_OPTION("no-shuffle", bool);
     SET_OPTION("tempdir", std::string);
-
+    SET_OPTION("sqlite", bool);
+  
+    
     SET_OPTION("optimizer", std::string);
     SET_OPTION_NONDEFAULT("optimizer-params", std::vector<float>);
     SET_OPTION("optimizer-delay", size_t);