+++ src/data/corpus.cpp
@@ -12,8 +12,8 @@ typedef std::vector<WordMask> SentBatch;
 
 CorpusIterator::CorpusIterator() : pos_(-1), tup_(0) {}
 
-CorpusIterator::CorpusIterator(Corpus& corpus)
-    : corpus_(&corpus), pos_(0), tup_(corpus_->next()) {}
+CorpusIterator::CorpusIterator(CorpusBase* corpus)
+    : corpus_(corpus), pos_(0), tup_(corpus_->next()) {}
 
 void CorpusIterator::increment() {
   tup_ = corpus_->next();
@@ -134,7 +134,7 @@ Corpus::Corpus(std::vector<std::string> paths,
                std::vector<Ptr<Vocab>> vocabs,
                Ptr<Config> options,
                size_t maxLength)
-    : DatasetBase(paths),
+    : CorpusBase(paths),
       options_(options),
       vocabs_(vocabs),
       maxLength_(maxLength ? maxLength : options_->get<size_t>("max-length")),